requirements.txt
altair==4.1.0
anyio==2.0.2
appdirs==1.4.3
argcomplete==1.12.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
awesome-streamlit
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
base58==2.1.0
bleach==3.2.2
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.1
catboost==0.24.4
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
confuse==1.4.0
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
eli5==0.11.0
entrypoints==0.3
fastapi==0.63.0
gitdb==4.0.5
GitPython==3.1.12
graphviz==0.16
h11==0.12.0
html5lib==1.0.1
htmlmin==0.1.12
httptools==0.1.1
idna==2.10
ImageHash==4.2.0
iniconfig==1.1.1
ipaddr==2.2.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.3.3
missingno==0.4.2
mistune==0.8.4
msgpack==0.6.2
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numba==0.52.0
numpy==1.19.5
packaging==20.8
pandas==1.2.1
pandas-profiling==2.10.0
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
PDPbox==0.2.0
pep517==0.8.2
pexpect==4.8.0
phik==0.11.0
pickleshare==0.7.5
Pillow==8.1.0
pipx==0.16.0.0
plotly==4.14.3
plotly-express==0.4.1
pluggy==0.13.1
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyarrow==2.0.0
pycparser==2.20
pydantic==1.7.3
pydeck==0.5.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
shap==0.38.1
six==1.15.0
slicer==0.0.7
smmap==3.0.5
sniffio==1.2.0
SQLAlchemy==1.3.22
starlette==0.13.6
statsmodels==0.12.1
streamlit==0.75.0
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.2
userpath==1.4.2
uvicorn==0.13.3
uvloop==0.14.0
validators==0.18.2
visions==0.6.0
watchdog==1.0.2
watchgod==0.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
xgboost==1.3.3
